+++ include/drawing.h
@@ -6,7 +6,7 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 20:37:11 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/11/13 00:35:06 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/11/13 01:00:26 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,6 @@
 
 # include "type.h"
 
-void	init_player(t_player *player);
-int		game_loop(void *arg);
-void	prepare_dda(t_world *world, t_dda *dda, int x);
+void	init_mlx_data(t_mlx_data *mlx_data);
 
 #endif
